backend/requirements.txt
# Core dependencies
anthropic==0.40.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
websockets==13.1

# Optional but recommended
python-multipart==0.0.12
aiofiles==24.1.0

# Development dependencies (optional)
pytest==8.3.0
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.0